+++ @vue-start/pro/src/comp/layout/index.ts
@@ -1,5 +1,7 @@
 export * from "./ctx";
 
+export * from "./Breadcrumb";
+
 export * from "./RouterView";
 export * from "./Layout";